requirements.txt
flask>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.0
pyarrow>=12.0.0
opencv-python>=4.8.0